Address PR #53 review feedback on env loaders and logging.
Remove unused headroom-ai, dedupe infra env parsing, wire Supabase loader
into smoke scripts, normalize contentType once in r2client, and gate full
allowedOrigins lists behind LOG_LEVEL=debug.
