+++ package.json
@@ -27,7 +27,6 @@
     "@types/react-syntax-highlighter": "^15.5.13",
     "framer-motion": "^12.34.3",
     "gsap": "^3.14.2",
-    "headroom-ai": "^0.22.4",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.575.0",
     "mermaid": "^11.14.0",
